daemon/requirements.txt
pyserial>=3.5
psutil>=5.9.0
mcp>=1.0.0
